+++ src/views/Intern/LoggInn/LoggInn.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState } from "react";
+import React, { FC } from "react";
 
 import { useAuth0 } from "@auth0/auth0-react";
